made clarifications abotu V3PA angle
There's a some confusing nomenclature regarding the V3PA angle needed to compute
the slew. The V3PA angle that is specified in the APT PA range special
requirement field is the PA angle of the V3 axis *as measured at the aperture's
reference position*, and corresponds to the ROLL_REF header keyword. This is not
the same as the PA_V3 header keyword, which corresponds to the PA angle of V3 at
the telescope boresight. These two angles are not the same because of **waves
hands** spherical trigonometry. We want to use the ROLL_REF/aperture V3 value.